pull all SQLite stuff into a single file
